feat: of/off 유효성 검증 처리
