--- docker-compose.yml
@@ -1,82 +1,0 @@
-version: '3.8'
-
-services:
-  pickcar-pub:
-    image: openjdk:21-jdk
-    container_name: pickcar-pub
-    ports:
-      - "8080:8080"
-    depends_on:
-      rabbitmq:
-        condition: service_healthy
-      mysql:
-        condition: service_healthy
-    volumes:
-      - ./emulator/build/libs/emulator-0.0.1-SNAPSHOT.jar:/app/app.jar
-    working_dir: /app
-    command: ["java", "-jar", "app.jar"]
-    networks:
-      - pickcar-net
-
-  pickcar-sub:
-    image: openjdk:21-jdk
-    container_name: pickcar-sub
-    ports:
-      - "8081:8080"
-    depends_on:
-      rabbitmq:
-        condition: service_healthy
-      mysql:
-        condition: service_healthy
-    volumes:
-      - ./consumer/build/libs/consumer-0.0.1-SNAPSHOT.jar:/app/app.jar
-    working_dir: /app
-    command: ["java", "-jar", "app.jar"]
-    networks:
-      - pickcar-net
-
-  rabbitmq:
-    image: rabbitmq:3-management
-    container_name: pickcar-rabbitmq
-    ports:
-      - "5672:5672"
-      - "15672:15672"
-    environment:
-      RABBITMQ_DEFAULT_USER: guest
-      RABBITMQ_DEFAULT_PASS: guest
-    networks:
-      - pickcar-net
-    healthcheck:
-      test: ["CMD", "rabbitmqctl", "status"]
-      interval: 10s
-      timeout: 5s
-      retries: 5
-
-  mysql:
-    image: mysql:8.0
-    container_name: pickcar-mysql
-    restart: always
-    environment:
-      MYSQL_ROOT_PASSWORD: pickcar
-      MYSQL_DATABASE: pickcar
-      MYSQL_USER: pickcar
-      MYSQL_PASSWORD: pickcar
-    ports:
-      - "3306:3306"
-    volumes:
-      - mysql-data:/var/lib/mysql
-    networks:
-      - pickcar-net
-    healthcheck:
-      test: [ "CMD", "mysqladmin", "ping", "-h", "localhost", "-p$MYSQL_ROOT_PASSWORD" ]
-      interval: 10s
-      timeout: 5s
-      retries: 5
-
-volumes:
-  mysql-data:
-    name: mysql-data
-
-networks:
-  pickcar-net:
-    name: pickcar-net
